requirements.txt
argon2-cffi==20.1.0
asgiref @ file:///tmp/build/80754af9/asgiref_1617740417847/work
bcrypt @ file:///C:/ci/bcrypt_1607022693089/work
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi @ file:///C:/ci/cffi_1613247431367/work
Django @ file:///tmp/build/80754af9/django_1620913015185/work
django-bootstrap-v5==1.0.4
django-ckeditor==6.1.0
django-js-asset==1.2.2
flake8 @ file:///tmp/build/80754af9/flake8_1620776156532/work
gunicorn==20.1.0
importlib-metadata @ file:///C:/ci/importlib-metadata_1617877477759/work
mccabe==0.6.1
misaka @ file:///D:/bld/misaka_1602586460495/work
olefile==0.46
Pillow==8.3.1
psycopg2 @ file:///C:/ci/psycopg2_1612298766762/work
pycodestyle @ file:///tmp/build/80754af9/pycodestyle_1615748559966/work
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pyflakes @ file:///tmp/build/80754af9/pyflakes_1617200973297/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
six @ file:///C:/ci/six_1605205404068/work
soupsieve==2.2.1
sqlparse @ file:///tmp/build/80754af9/sqlparse_1602184451250/work
text-unidecode==1.3
whitenoise @ file:///home/conda/feedstock_root/build_artifacts/whitenoise_1602596638218/work
wincertstore==0.2
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
